python/restApi/requirements.txt
ansicon==1.89.0
blessed==1.20.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
inquirer==3.1.4
itsdangerous==2.1.2
Jinja2==3.1.2
jinxed==1.2.0
MarkupSafe==2.1.3
python-dotenv==1.0.0
python-editor==1.0.4
readchar==4.0.5
six==1.16.0
wcwidth==0.2.12
Werkzeug==3.0.1
